+++ src/components/Flyout/FlyoutContent.tsx
@@ -11,7 +11,7 @@ import type * as T from "./Flyout.types";
 import s from "./Flyout.module.css";
 
 const FlyoutContent = (props: T.ContentProps) => {
-	const { children, className, attributes } = props;
+	const { children, className, attributes, flyoutAttributes, flyoutClassName } = props;
 	const {
 		flyout,
 		id,
@@ -92,8 +92,9 @@ const FlyoutContent = (props: T.ContentProps) => {
 
 	if (status === "idle" || !mounted) return null;
 
-	const contentClassNames = classNames(
+	const rootClassNames = classNames(
 		s.content,
+		flyoutClassName,
 		triggerType === "hover" && s["--hover"],
 		status === "visible" && s["--visible"],
 		// animating only when we're opening the first flyout or closing the last flyout within the same cooldown
@@ -124,12 +125,14 @@ const FlyoutContent = (props: T.ContentProps) => {
 		<ContentProvider value={{ elRef: flyoutElRef }}>
 			{/* eslint-disable-next-line jsx-a11y/no-static-element-interactions */}
 			<div
-				className={contentClassNames}
+				{...flyoutAttributes}
+				className={rootClassNames}
 				style={
 					{
+						...flyoutAttributes?.style,
 						...styles,
 						"--rs-flyout-gap": contentGap,
-					} as React.CSSProperties
+					} as unknown as React.CSSProperties
 				}
 				ref={flyoutElRef}
 				onTransitionEnd={handleTransitionEnd}