+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 3.6.0
 
 - Flyout: Exported from the library
+- Flyout: Moved css to css layers
+- Flyout: flyoutClassName, flyoutAttributes?